Workaround unused-but-set-variable warning.
The UART status register must be read to clear the flags and enable operation.
Previously, it was read into a dummy variable which triggered the warning.
Attempts to suppress the warning didn't work for some reason, but the
statement "UART0_SR;" is sufficient to emit assembly to read the register.

I don't trust this to always work, so there is a comment with the purpose
of the line.